#37 - Feat: JWT AccessToken 화이트 리스트 방식 적용(2차 화이트리스트 검증)

+++ 4Week_Mission/mutbooks/src/main/java/com/example/mutbooks/app/security/jwt/JwtProvider.java
@@ -25,7 +25,7 @@ public class JwtProvider {
     }
 
     // JWT Access Token 발급
-    public String generateAccessToken(Map<String, Object> claims, int seconds) {
+    public String generateAccessToken(Map<String, Object> claims, long seconds) {
         long now = new Date().getTime();
         Date accessTokenExpiresIn = new Date(now + 1000L * seconds);
